updated cacheTests, +CommandExe unit tests
