+++ EasyKeys.Extensions.Dapper.UnitTest/CommandExecuter/CommandExecuterTests.cs
@@ -1,0 +1,6 @@
+﻿namespace EasyKeys.Extensions.Dapper.UnitTest.CommandExe
+{
+    public class CommandExecuterTests
+    {
+    }
+}
